Update testing.
This updates and adds correct testing.
