+++ app/_sections/Navbar/_NavbarType3.tsx
@@ -1,5 +1,5 @@
 "use client";
-
+import React from "react";
 import { useEffect, useState } from "react";
 import { Dialog, Popover } from "@headlessui/react";
 import {
