--- tests/mockSupabaseClient.ts
@@ -1,28 +1,0 @@
-import { vi } from "vitest";
-
-export const mockSupabaseClient = {
-  auth: {
-    signInWithPassword: vi.fn(),
-    signUp: vi.fn(),
-    signOut: vi.fn(),
-    getSession: vi.fn(),
-  },
-  from: vi.fn((table: string) => ({
-    select: vi.fn().mockResolvedValue({ data: [], error: null }),
-    insert: vi.fn().mockResolvedValue({ data: null, error: null }),
-    update: vi.fn().mockResolvedValue({ data: null, error: null }),
-    delete: vi.fn().mockResolvedValue({ data: null, error: null }),
-    eq: vi.fn().mockReturnThis(),             
-    single: vi.fn().mockResolvedValue({ data: null, error: null }),
-    order: vi.fn().mockReturnThis(),
-  })),
-};
-
-// Mock `createClient` globally
-vi.mock("@supabase/supabase-js", async () => {
-  const actual = await vi.importActual("@supabase/supabase-js");
-  return {
-    ...actual,
-    createClient: vi.fn(() => mockSupabaseClient),
-  };
-});
