+++ vitest.config.ts
@@ -30,14 +30,10 @@ export default defineConfig({
       reporter: ["text", "html"],
       include: [
         "app/page.tsx",
-        "app/viewer/page.tsx",
-        "app/viewer/%5BprojectId%5D/%5BpageTitle%5D/page.tsx",
-        "app/viewer/%5BprojectId%5D/page.tsx",
-        "app/viewer/**/page.tsx",
-        "app/viewer/[projectId]/[pageTitle]/page.tsx",
-        "app/viewer/[projectId]/page.tsx",
-        "app/viewer/*/page.tsx", 
-        "app/viewer/*/*/page.tsx",
+        "app/%5BpageTitle%5D/page.tsx",
+        "app/[pageTitle]/page.tsx",
+        "app/*/page.tsx", 
+        "app/_sections/**/*.tsx",
         "app/_sections/Button.tsx",
         "app/_sections/ModalVideo.tsx",
         "app/_sections/CallToAction/_CallToActionType1.tsx",
